# application/api-spec/src/main/resources/openapi/food-rescue.yaml
openapi: 3.0.3
info:
  title: Savr Platform API
  version: 0.0.1-SNAPSHOT
  description: |
    REST API for the Savr platform.
servers:
  - url: http://localhost:8080
...
paths:
  /users:
    post:
      tags:
        - User
      summary: Create a user
      description: |
        Creates a new user in Keycloak and in the local database.

      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: User creation payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            examples:
              adminUser:
                summary: Admin user
                value:
                  firstName: Admin
                  lastName: User
                  email: admin@posam.sk
                  password: admin123
                  role: ADMIN
              regularUser:
                summary: Regular user
                value:
                  firstName: Jana
                  lastName: Novakova
                  email: jana.novakova@posam.sk
                  password: user123
                  role: USER
      responses:
        '201':
          description: User created successfully
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: Admin privileges required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                forbidden:
                  value:
                    code: FORBIDDEN
                    message: Access denied
                    details:
                      - Insufficient permissions
                    timestamp: 2025-04-08T18:30:00Z
                    path: /users
        '409':
          description: A user with the same email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicateUser:
                  value:
                    code: CONFLICT
                    message: User with email=jana.novakova@posam.sk already exists
                    details: []
                    timestamp: 2025-04-08T18:30:00Z
                    path: /users
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /users/register:
    post:
      tags:
        - User
      summary: Register a user
      description: |
        Registers a new regular user in Keycloak and in the local database.

      operationId: registerUser
      requestBody:
        required: true
        description: Public user registration payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
            examples:
              userRegistration:
                summary: Regular user registration
                value:
                  firstName: Jana
                  lastName: Novakova
                  email: jana.novakova@posam.sk
                  password: user123
      responses:
        '201':
          description: User registered successfully
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '409':
          description: A user with the same email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /reservations:
    get:
      tags:
        - Reservation
      summary: List reservations
      description: |
        Returns reservation history of the authenticated user, or reservations of a managed offer when `offerId` is provided.

      operationId: getReservations
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OfferIdQueryParam'
      responses:
        '200':
          description: Reservations listed successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to access the requested reservations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'

    post:
      tags:
        - Reservation
      summary: Create a reservation
      description: |
        Reserves an available offer for the authenticated user.

      operationId: createReservation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: Reservation creation payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
      responses:
        '201':
          description: Reservation created successfully
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to reserve this offer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Offer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Offer is no longer available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /reservations/{id}:
    get:
      tags:
        - Reservation
      summary: Get reservation by id
      description: |
        Returns reservation detail for its owner, the managed business, or admin.

      operationId: getReservation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReservationIdPathParam'
      responses:
        '200':
          description: Reservation found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to access this reservation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /reservations/{id}/cancel:
    post:
      tags:
        - Reservation
      summary: Cancel reservation
      description: |
        Cancels an active reservation before the pickup window ends.

      operationId: cancelReservation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReservationIdPathParam'
      responses:
        '200':
          description: Reservation cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to cancel this reservation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Reservation can no longer be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /reservations/{id}/pickup:
    post:
      tags:
        - Reservation
      summary: Confirm reservation pickup
      description: |
        Confirms pickup of a reserved offer by the business owner or admin.

      operationId: confirmReservationPickup
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReservationIdPathParam'
      responses:
        '200':
          description: Reservation pickup confirmed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to confirm pickup for this reservation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OfferIdQueryParam:
      name: offerId
      in: query
      required: false
      description: Optional offer identifier used to list reservations of a managed offer
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 100
...
    ReservationIdPathParam:
      name: id
      in: path
      required: true
      description: Reservation identifier
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 500
...
  responses:
    BadRequestResponse:
      description: Invalid request body or request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidBody:
              value:
                code: VALIDATION_ERROR
                message: Invalid request body
                details:
                  - name must not be blank
                timestamp: 2025-04-08T18:30:00Z
                path: /businesses
    UnauthorizedResponse:
      description: Authentication required or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                code: UNAUTHORIZED
                message: Full authentication is required to access this resource
                details: []
                timestamp: 2025-04-08T18:30:00Z
                path: /businesses
    InternalServerErrorResponse:
      description: Unexpected internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internalError:
              value:
                code: INTERNAL_ERROR
                message: Unexpected internal error
                details:
                  - Please contact support
                timestamp: 2025-04-08T18:30:00Z
                path: /businesses
  schemas:
    CreateUserRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - password
        - role
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          example: Admin
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          example: User
        email:
          type: string
          format: email
          maxLength: 255
          example: admin@posam.sk
        password:
          type: string
          format: password
          minLength: 6
          maxLength: 255
          example: admin123
        role:
          $ref: '#/components/schemas/UserRole'
    RegisterUserRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - password
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          example: Jana
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          example: Novakova
        email:
          type: string
          format: email
          maxLength: 255
          example: jana.novakova@posam.sk
        password:
          type: string
          format: password
          minLength: 6
          maxLength: 255
          example: user123
...
    CreateReservationRequest:
      type: object
      required:
        - offerId
        - quantity
      properties:
        offerId:
          type: integer
          format: int64
          minimum: 1
          example: 100
        quantity:
          type: integer
          minimum: 1
          example: 1
    ReservationResponse:
      type: object
      required:
        - id
        - offerId
        - userId
        - quantity
        - status
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          example: 500
        offerId:
          type: integer
          format: int64
          example: 100
        userId:
          type: integer
          format: int64
          example: 2
        quantity:
          type: integer
          example: 1
        status:
          $ref: '#/components/schemas/ReservationStatus'
        createdAt:
          type: string
          format: date-time
          example: 2025-04-08T16:30:00Z
        cancelledAt:
          type: string
          format: date-time
          nullable: true
          example: 2025-04-08T17:00:00Z
        pickupConfirmation:
          $ref: '#/components/schemas/PickupConfirmation'
    PickupConfirmation:
      type: object
      required:
        - confirmedByUserId
        - confirmedAt
      properties:
        confirmedByUserId:
          type: integer
          format: int64
          example: 10
        confirmedAt:
          type: string
          format: date-time
          example: 2025-04-08T19:10:00Z
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - details
        - timestamp
        - path
      properties:
        code:
          type: string
          example: CONFLICT
        message:
          type: string
          example: Business with the same name already exists for this owner
        details:
          type: array
          items:
            type: string
          example:
            - Duplicate or invalid database state
        timestamp:
          type: string
          format: date-time
          example: 2025-04-08T18:30:00Z
        path:
          type: string
          example: /businesses
    UserRole:
      type: string
      enum:
        - USER
        - ADMIN
...
    ReservationStatus:
      type: string
      enum:
        - ACTIVE
        - CANCELLED
        - PICKED_UP
        - NO_SHOW
...